refactor: 전역 상태 파일명 변경
